+++ lib/src/default_index/readonly.rs
@@ -25,11 +25,10 @@ use std::sync::Arc;
 use byteorder::{LittleEndian, ReadBytesExt};
 use smallvec::SmallVec;
 
-use super::composite::CompositeIndex;
+use super::composite::{CompositeIndex, IndexSegment};
 use super::entry::{IndexEntry, IndexPosition, SmallIndexPositionsVec};
 use super::mutable::DefaultMutableIndex;
 use super::store::IndexLoadError;
-use super::IndexSegment;
 use crate::backend::{ChangeId, CommitId, ObjectId};
 use crate::index::{HexPrefix, Index, MutableIndex, PrefixResolution, ReadonlyIndex};
 use crate::revset::{ResolvedExpression, Revset, RevsetEvaluationError};
